replay 400 errors
the lemmy backend server for some reason returns a 400 bad request sometimes early on when it cannot handle a request. This is obviously not the point of a 400 client error. So we will retry every 400 error at least once, very inefficient but likely not to be a problem unless we get a high volume of bad requests